Update packageManager
Automatic selection of emulators if roms are available
